@@ -1,6 +1,33 @@
+[tool.poetry]
+name = "aws-log-ingestion"
+version = "2.7.1"
+description = ""
+authors = ["New Relic <serverless@newrelic.com>"]
+license = "Apache 2.0"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+aiohttp = "^3.8.4"
+
+[tool.poetry.group.dev.dependencies]
+asynctest = "^0.13.0"
+aws-sam-cli = "^1.33.0"
+black = "^23.3.0"
+boto3 = "^1.19.2"
+coverage = "^6.0.2"
+flake8 = "^4.0.1"
+mock = "^4.0.3"
+pytest = "^6.2.5"
+pytest-asyncio = "^0.16.0"
+serverlessrepo = {git = "https://github.com/kolanos/aws-serverlessrepo-python.git", rev = "master"}
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
+
 [tool.black]
 line-length = 88
-target-version = ['py37']
+target-version = ['py39']
 include = '\.pyi?$'
 exclude = '''
 /(
